Make LR_AUTH_ONLY enum value language compliant
GCC with -wpedantic warns:

/home/test/librepo/librepo/types.h:77:27: warning: enumerator value for ‘LR_AUTH_ONLY’ is not an intege
r constant expression [-Wpedantic]
   77 |     LR_AUTH_ONLY        = (1<<31), /*!< This is a meta symbol. OR this value
      |                           ^

That's because enum type is equaled to signed int (in C99 this code
uses for compilation) and 1<<31 does not fit into 32-bit signed
integer (which is a size of int on Linux GCC).

Semantically correct fix would be changing it into 1<<30 or MAX_INT,
but that would break ABI.

Therefore this patch retains the value and fixes the warning by
casting the desired bit-value to an unsigned int explicitly.

+++ librepo/types.h
@@ -74,7 +74,7 @@ typedef enum {
     LR_AUTH_NTLM        = (1<<3),  /*!< HTTP NTLM authentication */
     LR_AUTH_DIGEST_IE   = (1<<4),  /*!< HTTP Digest authentication with an IE flavor */
     LR_AUTH_NTLM_WB     = (1<<5),  /*!< NTLM delegating to winbind helper */
-    LR_AUTH_ONLY        = (1<<31), /*!< This is a meta symbol. OR this value
+    LR_AUTH_ONLY        = ((int)(1u<<31)), /*!< This is a meta symbol. OR this value
                                         together with a single specific auth
                                         value to force libcurl to probe for
                                         un-restricted auth and if not, only
